+++ oap-server/server-query-plugin/promql-plugin/src/main/java/org/apache/skywalking/oap/query/promql/rt/PromQLExprQueryVisitor.java
@@ -440,6 +440,32 @@ public class PromQLExprQueryVisitor extends PromQLParserBaseVisitor<ParseResult>
                 checkLabels(labelMap, LabelName.ENDPOINT);
                 entity.setEndpointName(labelMap.get(LabelName.ENDPOINT));
                 break;
+            case ServiceRelation:
+                checkLabels(labelMap, LabelName.DEST_SERVICE);
+                checkLabels(labelMap, LabelName.DEST_LAYER);
+                entity.setDestServiceName(labelMap.get(LabelName.DEST_SERVICE));
+                entity.setDestNormal(Layer.nameOf(labelMap.get(LabelName.DEST_LAYER)).isNormal());
+                break;
+            case ServiceInstanceRelation:
+                checkLabels(labelMap, LabelName.DEST_SERVICE);
+                checkLabels(labelMap, LabelName.DEST_SERVICE_INSTANCE);
+                checkLabels(labelMap, LabelName.SERVICE_INSTANCE);
+                checkLabels(labelMap, LabelName.DEST_LAYER);
+                entity.setServiceInstanceName(labelMap.get(LabelName.SERVICE_INSTANCE));
+                entity.setDestServiceInstanceName(labelMap.get(LabelName.DEST_SERVICE_INSTANCE));
+                entity.setDestServiceName(labelMap.get(LabelName.DEST_SERVICE));
+                entity.setDestNormal(Layer.nameOf(labelMap.get(LabelName.DEST_LAYER)).isNormal());
+                break;
+            case EndpointRelation:
+                checkLabels(labelMap, LabelName.DEST_SERVICE);
+                checkLabels(labelMap, LabelName.DEST_ENDPOINT);
+                checkLabels(labelMap, LabelName.ENDPOINT);
+                checkLabels(labelMap, LabelName.DEST_LAYER);
+                entity.setEndpointName(labelMap.get(LabelName.ENDPOINT));
+                entity.setDestEndpointName(labelMap.get(LabelName.DEST_ENDPOINT));
+                entity.setDestServiceName(labelMap.get(LabelName.DEST_SERVICE));
+                entity.setDestNormal(Layer.nameOf(labelMap.get(LabelName.DEST_LAYER)).isNormal());
+                break;
         }
         return entity;
     }