+++ docs/en/changes/changes.md
@@ -6,6 +6,7 @@
 #### OAP Server
 * Fix wrong indices in the eBPF Profiling related models.
 * Support exclude the specific namespaces traffic in the eBPF Access Log receiver.
+* Support to query relation metrics through PromQL.
 
 
 #### UI